working on tex_box.py
